Support multiple image updates in single PR (#7)
* Support multiple image updates in one action

This is a breaking change so will be part of a v2 release. The `path` and `image-name` inputs have been replaced by a single `images` input that takes a multiline, comma delimited string containing the directory path and image name per line. This will allow us to create a single PR for multiple image tag updates from the source repo.

* Update project description

* Update action docs

* Update example for v2