Set kustomize-version to 5.1.1 due to error
There is an error with the archive format as part of the 5.2.0 release that is causing all our deployment builds to fail.

https://github.com/kubernetes-sigs/kustomize/issues/5396

For now just add an optional variable so we can also control this at the organization or repository level. We can make this an action input as well.